--- a/java/src/test/java/com/example/.gitignore
+++ b/java/src/test/java/com/example/.gitignore
@@ -1,0 +1,1 @@
+gentest