Fixes https://github.com/splunk/stef/issues/93
Fixes https://github.com/splunk/stef/issues/95

### Overview

The specification did not properly define how codecs for recursive
types were expected to work. The implementation also contained bugs.

I added several tests that verify schemas with recursive types, see
TestGenerate(). The test generates Go and Java code and runs the tests.

These tests have found numerous bugs in codecs when recursing definitions
were present. All found bugs are now fixed. There are likely more cases
to be tested, particularly when recursion involves oneof or struct only.
I am postponing that to a future PR since this PR is already too big.

There was no good way to split this PR into smaller ones since fixing
one codec required touching other codecs unfortunately since recursion
is involved here.

### What's changed

Codecs now detect recursion at runtime. Codecs keep a separate "last value"
state at each recursion level and apply differential encoding at each level
independently. A stack of such "last values" is maintained by array and
multimap codecs. See specification.md for further explanation.

Previously the recursion was detected statically, at schema compile time.
This was an incorrect approach since the node where detection must happen
depends on the traversal order of the schema, and the traversal order
depends on which root struct we start from. Since multiple roots are allowed
in the same schema, for each root the recursion detection may happen at a
different node. This means there cannot be a single statically defined
recursion detection point per schema, it needs to be dynamic, depending on
what root struct is being encoded.

I will need to further cleanup the code that is responsible for compile-time
recursion detection. I haven't decided yet if I want to completely remove it
or keep it for diagnostic purposes and to output more meaningful error
messages during schema compilation. This will be done in a future PR.

### Performance

The change initially resulted in significant performance degradation,
encoding was around 20% slower. This was unacceptable, and I was forced to
perform some optimizations, including in adjacent code to claw back the
performance lost and bring it back to within 1% of previous performance.

Here are benchmark diffs:
```
                               │ bench_main.txt │         bench_current.txt          │
                               │     sec/op     │   sec/op     vs base               │
SerializeNative/STEF/none-10        4.353m ± 1%   4.322m ± 1%  -0.72% (p=0.033 n=15)
DeserializeNative/STEF/none-10      1.702m ± 0%   1.703m ± 1%       ~ (p=0.486 n=15)
geomean                             2.722m        2.713m       -0.31%

                               │ bench_main.txt │         bench_current.txt          │
                               │   sec/point    │  sec/point   vs base               │
SerializeNative/STEF/none-10        65.11n ± 1%   64.65n ± 1%  -0.71% (p=0.033 n=15)
DeserializeNative/STEF/none-10      25.45n ± 0%   25.48n ± 1%       ~ (p=0.505 n=15)
geomean                             40.71n        40.59n       -0.30%

                               │ bench_main.txt │          bench_current.txt          │
                               │      B/op      │     B/op      vs base               │
SerializeNative/STEF/none-10       3.516Mi ± 0%   3.517Mi ± 0%       ~ (p=0.935 n=15)
DeserializeNative/STEF/none-10     839.3Ki ± 0%   840.7Ki ± 0%  +0.17% (p=0.000 n=15)
geomean                            1.698Mi        1.699Mi       +0.10%

                               │ bench_main.txt │         bench_current.txt          │
                               │   allocs/op    │  allocs/op   vs base               │
SerializeNative/STEF/none-10        2.803k ± 0%   2.860k ± 0%  +2.03% (p=0.000 n=15)
DeserializeNative/STEF/none-10      1.260k ± 0%   1.318k ± 0%  +4.60% (p=0.000 n=15)
geomean                             1.879k        1.942k       +3.31%
```

### Future Work

Generally speaking more testing work is necessary to ensure codecs work.
This will be done in future PRs.